Update draft submission for open source hardware (#121)
* Update draft submission for open source hardware

Following input from the Commission, this updated draft focuses solely on open source hardware source files and doesn't introduce the term of open source hardware in the submission to avoid confusion.

It also clarifies that open source hardware source files are outside of the regulation unless they are 'made available on the market' in the course of a commercial activity.